kprobes: cheri adaptation
Adapt the code to register, execute and unregister kprobes for cheri
systems.

When a kprobe is registered, assume that kprobe->addr contains just the
address of the probe point. Mostly, this info comes from userspace.
There's no way for the caller to set proper bounds and permissions to
install the probe at a kernel address.

Create non-executable data capabilities for the probe point and the slot
for the moved instruction. We can't produce a capability that's writable
and executable.

Try to set tight bounds for the slot capability. We only have two
instruction in the slot: The moved instruction from the probe point and
an uncompressed ebreak. Replace the current epc's address and bounds
with the slot capability to execute the two instructions.

For instructions that have to be simulated (because of pc-relative fields),
pass just the address to the simulation functions. We have to fabricate
or derive capabilities that match the behaviour of the simulated
instruction.

When a kprobe is registered, we have to create a "restore capability"
that is unsealed, executable and points to the instruction that follows
the probe point. We don't have access to the current epc at this point.
So we have to byte the bullet and derive the restore capability directly
from kernel_code_cap.

The adapted kprobes code passes all applicable kprobes selftests. There's
kunit tests to verify that the simulated instructions behave identical to
native instructions. These tests are also passing.

+++ include/linux/kprobes.h
@@ -400,7 +400,7 @@ static inline struct kprobe_ctlblk *get_kprobe_ctlblk(void)
 }
 
 kprobe_opcode_t *kprobe_lookup_name(const char *name, unsigned int offset);
-kprobe_opcode_t *arch_adjust_kprobe_addr(unsigned long addr, unsigned long offset, bool *on_func_entry);
+kprobe_opcode_t *arch_adjust_kprobe_addr(uintptr_t addr, unsigned long offset, bool *on_func_entry);
 
 int register_kprobe(struct kprobe *p);
 void unregister_kprobe(struct kprobe *p);
